+++ SRC/Public/Abstractions/AbstractServiceEntry.cs
@@ -25,8 +25,9 @@ namespace Solti.Utils.DI
         /// </summary>
         /// <param name="interface">The interface of the service.</param>
         /// <param name="name">The (optional) name of the service.</param>
+        /// <param name="owner">The owner of this entry.</param>
         /// <exception cref="ArgumentException">The <paramref name="interface"/> is not an interface.</exception>
-        public AbstractServiceEntry(Type @interface, string? name): this(@interface, name, null, null)
+        public AbstractServiceEntry(Type @interface, string? name, IServiceContainer owner): this(@interface, name, null, owner)
         { 
         }
 
@@ -36,12 +37,13 @@ namespace Solti.Utils.DI
         /// <param name="interface">The interface of the service.</param>
         /// <param name="name">The (optional) name of the service.</param>
         /// <param name="lifetime">The (optional) lifetime of the service.</param>
-        /// <param name="owner">The (optional) owner of this entry.</param>
+        /// <param name="owner">The owner of this entry.</param>
         /// <exception cref="ArgumentException">The <paramref name="interface"/> is not an interface.</exception>
-        internal protected AbstractServiceEntry(Type @interface, string? name, Lifetime? lifetime, IServiceContainer? owner)
+        protected AbstractServiceEntry(Type @interface, string? name, Lifetime? lifetime, IServiceContainer owner)
         {
             Ensure.Parameter.IsNotNull(@interface, nameof(@interface));
             Ensure.Parameter.IsInterface(@interface, nameof(@interface));
+            Ensure.Parameter.IsNotNull(owner, nameof(owner));
 
             Interface = @interface;
             Name      = name;
@@ -68,9 +70,9 @@ namespace Solti.Utils.DI
         public Lifetime? Lifetime { get; }
 
         /// <summary>
-        /// The (optional) owner of this entry.
+        /// The owner of this entry.
         /// </summary>
-        public IServiceContainer? Owner { get; }
+        public IServiceContainer Owner { get; }
 
         /// <summary>
         /// The (optional) implementation of the service.